Forced BigDecimal type for money handling rather than double, implemented ExpenseManager.java, going to work on developing test cases for ExpenseManager.java inside of ExpenseManagerTests.java

+++ tests/model/RoommateTests.java
@@ -17,22 +17,22 @@ public class RoommateTests {
     @Test
     public void testUpdateBalancePositive(){
         Roommate roommate = new Roommate("Nick", BigDecimal.ZERO);
-        roommate.updateBalance(9.99);
+        roommate.updateBalance(BigDecimal.valueOf(9.99));
         assertEquals(BigDecimal.valueOf(9.99), roommate.getBalance());
     }
 
     @Test
     public void testUpdateBalanceNegative(){
         Roommate roommate = new Roommate("David", BigDecimal.ZERO);
-        roommate.updateBalance(-3.14);
+        roommate.updateBalance(BigDecimal.valueOf(-3.14));
         assertEquals(BigDecimal.valueOf(-3.14), roommate.getBalance());
     }
 
     @Test
     public void testResetBalance(){
         Roommate roommate = new Roommate("Zachary", BigDecimal.ZERO);
-        roommate.updateBalance(74.99);
-        roommate.updateBalance(-14.99);
+        roommate.updateBalance(BigDecimal.valueOf(74.99));
+        roommate.updateBalance(BigDecimal.valueOf(-14.99));
         roommate.resetBalance();
         assertEquals(BigDecimal.ZERO, roommate.getBalance());
     }
